add feature code/encode image to/from text base64
