+++ DTAdmin/Models/DataManager.swift
@@ -93,9 +93,11 @@ class DataManager: HTTPManager {
                     return
                 }
                 DispatchQueue.main.async {
+                    
                     if responseValue.statusCode == HTTPStatusCodes.OK.rawValue {
                         completionHandler(json, nil)
-                    } else {
+                    }
+                    else {
                         var errorMsg: String = ""
                         if let errorReason = json as? [String: String]  {
                             guard let errorServerMsg = errorReason["response"] else { return }
